# openapi/backend/openapi.yaml
...
openapi: '3.0.4'
info:
  title: webstatus.dev API
  version: 0.1.0
  description: >
    A tool to monitor and track the status of all Web Platform features across
    dimensions that are related to availability and implementation quality
    across browsers, and adoption by web developers.
servers:
  - url: https://api.webstatus.dev
    description: Production
paths:
...
  /v1/stats/features/browsers/{browser}/missing_one_implementation_counts:
    parameters:
      - $ref: '#/components/parameters/browserPathParam'
    get:
      summary: >
        Returns the count of features that are supported by all of the specified
        comparison browsers but not supported by the specified target browser,
        as of each browser's (target browser or comparison browser) release
        date.
      operationId: listMissingOneImplementationCounts
      parameters:
        - $ref: '#/components/parameters/startAtParam'
        - $ref: '#/components/parameters/endAtParam'
        - $ref: '#/components/parameters/paginationTokenParam'
        - $ref: '#/components/parameters/paginationSizeParam'
        - $ref: '#/components/parameters/includeBaselineMobileBrowsersParam'
        - in: query
          name: browser
          description: >
            A list of browsers to check if {browser} is lagging behind on
            features. The list is provided by specifying repeating query parameters.
            Example: ?browser=fooBrowser&browser=bazBrowser
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SupportedBrowsers'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserReleaseFeatureMetricsPage'
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '429':
          description: Rate Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
...
  /v1/saved-searches/{search_id}:
    parameters:
      - name: search_id
        in: path
        description: Saved Search ID
        required: true
        schema:
          type: string
    get:
      summary: Get saved search
      operationId: getSavedSearch
      security:
        - bearerAuth: []
        # User can opt out of providing a bearer token.
        # They just won't get bookmark and role information.
        - noAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearchResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
    patch:
      operationId: updateSavedSearch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearchResponse'
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
    delete:
      operationId: removeSavedSearch
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content (successful deletion)
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found (saved search does not exist)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
...
components:
  parameters:
    browserPathParam:
      in: path
      name: browser
      description: Browser name
      required: true
      schema:
        $ref: '#/components/schemas/SupportedBrowsers'
...
    startAtParam:
      in: query
      name: startAt
      schema:
        type: string
        format: date
      description: Start date (RFC 3339, section 5.6, for example, 2017-07-21). The date is inclusive.
      required: true
    # TODO: Create a end_at version of this parameter
    endAtParam:
      in: query
      name: endAt
      schema:
        type: string
        format: date
      description: End date (RFC 3339, section 5.6, for example, 2017-07-21). The date is exclusive.
      required: true
    paginationTokenParam:
      in: query
      name: page_token
      schema:
        type: string
      description: Pagination token
      required: false
    paginationSizeParam:
      in: query
      name: page_size
      schema:
        type: integer
        maximum: 100
        minimum: 1
        default: 1
      required: false
      description: Number of results to return
    includeBaselineMobileBrowsersParam:
      in: query
      name: include_baseline_mobile_browsers
      schema:
        type: boolean
      description: >
        Include baseline mobile browsers in the response. This is only applicable
        for desktop browsers with a corresponding mobile browser.
      required: false
  schemas:
    SupportedBrowsers:
      type: string
      description: List of supported browsers that webstatus.dev currently ingests
      enum:
        - chrome
        - firefox
        - safari
        - edge
        - chrome_android
        - firefox_android
        - safari_ios
...
    BrowserReleaseFeatureMetric:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: >
            The timestamp that correlates to an event that may influence the
            count of features for a browser. This may be the release of the
            browser itself (when a browser may support a new set of features),
            or the release of another browser (when another browser supports
            a feature but our specified browser is now lagging behind with a new
            feature). Refer to the individual endpoint for more context on the
            use of this component.
        count:
          type: integer
          description: Total count of features.
          format: int64
      required:
        # For now, only require timestamp in case the definition
        # of the metric needs to change. Similar to WPTRunMetric.
        - timestamp
    BrowserReleaseFeatureMetricsPage:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/PageMetadata'
        data:
          type: array
          items:
            $ref: '#/components/schemas/BrowserReleaseFeatureMetric'
      required:
        - data
...
    PageMetadata:
      type: object
      properties:
        next_page_token:
          type: string
...
    GenericUpdatableUniqueModel:
      type: object
      properties:
        id:
          type: string
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required:
        - id
        - created_at
        - updated_at
    SavedSearch:
      type: object
      description: |
        Represents a saved search with a name and a query.
        Both `name` and `query` are required when creating a new saved search.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          minLength: 1
          maxLength: 1024
        query:
          type: string
          minLength: 1
          maxLength: 256
      required:
        - name
        - query
    SavedSearchUpdateRequest:
      type: object
      description: |
        Represents the data required to update a saved search.
        Only the fields that need to be modified need to be included in the request.
        To clear a field, include the field in `update_mask` and set the corresponding value to null.
        Only fields with nullable set to true can be cleared.
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
          minLength: 1
        query:
          type: string
          minLength: 1
        update_mask:
          type: array
          description: >
            A list of fields to update. Required. Allowed values are: `name`, `description`, `query`.
          items:
            type: string
            enum:
              - name
              - description
              - query
            # Custom field used by https://github.com/oapi-codegen/oapi-codegen
            # Otherwise, the constant name will be Name, Description and Query
            x-enumNames:
              - SavedSearchUpdateRequestMaskName
              - SavedSearchUpdateRequestMaskDescription
              - SavedSearchUpdateRequestMaskQuery
          minItems: 1
          uniqueItems: true
      required:
        - update_mask
    UserSavedSearchBookmarkStatus:
      type: string
      description: |
        The bookmark status for a saved search for a user.
      enum:
        - bookmark_none
        - bookmark_active
    UserSavedSearchBookmark:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/UserSavedSearchBookmarkStatus'
      required:
        - status
    SavedSearchResponse:
      allOf:
        - $ref: '#/components/schemas/GenericUpdatableUniqueModel'
        - $ref: '#/components/schemas/SavedSearch'
        - type: object
          properties:
            permissions:
              $ref: '#/components/schemas/UserSavedSearchPermissions'
            bookmark_status:
              $ref: '#/components/schemas/UserSavedSearchBookmark'
    UserSavedSearchRole:
      type: string
      enum:
        - saved_search_owner
    UserSavedSearchPermissions:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/UserSavedSearchRole'
...
    BasicErrorModel:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
        code:
          type: integer
          minimum: 100
          maximum: 600
    ExtendedErrorModel:
      allOf: # Combines the BasicErrorModel and the inline model
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          required:
            - errors
          properties:
            errors:
              type: object
              additionalProperties:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    noAuth:
      description: No authentication required. User can opt out of providing a bearer token.
      type: apiKey
      in: header
      name: X-No-Auth